[dmbus] PV guest do not have dmbus connection
dmbus client handler was used to test for domain initialisation prior
OXT-544, this is no longer the case since PV-guest will set up vkbd
directly without talking to dm-agent.
Assumption was /client/ field in the /struct domains/ was always a valid
pointer for an initialized domains, but this is no longer the case.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-570