input: adjust limits for touchpad quirk.
Some touchpads report values outside of the boundaries they pretend to
honor. That has input-server consider more of the touchpad surface to be
"edge" than it should be.
This should probably be a per-device based quirk, but there is no easy
way to reliably do that in the current state of the daemon.

Instead, since the limits of the touchpad are cached in a struct, adjust
them if a value "reasonably" (~10% outside of the reported limits)
outside of the reported boundaries is reported.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-937